# optimize ci speed in expensive eager test