--- a/python/oneflow/test/expensive/test_convtranspose.py
+++ b/python/oneflow/test/expensive/test_convtranspose.py
@@ -278,7 +278,7 @@ class TestConvTranspose(flow.unittest.TestCase):
         for arg in GenArgList(arg_dict):
             arg[0](test_case, *arg[1:])
 
-    @autotest()
+    @autotest(n=5)
     def test_ConvTranspose1d_(test_case):
         channels = random(1, 6)
         m = torch.nn.ConvTranspose1d(
@@ -299,7 +299,7 @@ class TestConvTranspose(flow.unittest.TestCase):
         return y
 
     @unittest.skipIf(os.getenv("ONEFLOW_TEST_CPU_ONLY"), "only test cpu cases")
-    @autotest(n=30)
+    @autotest(n=5)
     def test_deconv1d_group_with_random_data(test_case):
         channels = 720  # lcm(1, 2, 3, 4, 5, 6)
         m = torch.nn.ConvTranspose1d(
@@ -322,7 +322,7 @@ class TestConvTranspose(flow.unittest.TestCase):
         y = m(x)
         return y
 
-    @autotest()
+    @autotest(n=5)
     def test_ConvTranspose3d_(test_case):
         channels = random(1, 2)
         m = torch.nn.ConvTranspose3d(
@@ -343,9 +343,9 @@ class TestConvTranspose(flow.unittest.TestCase):
         return y
 
     @unittest.skipIf(os.getenv("ONEFLOW_TEST_CPU_ONLY"), "only test cpu cases")
-    @autotest(n=15)
+    @autotest(n=5)
     def test_deconv3d_group_with_random_data(test_case):
-        channels = 720  # lcm(1, 2, 3, 4, 5, 6)
+        channels = 120  # lcm(1, 2, 3, 4, 5)
         m = torch.nn.ConvTranspose3d(
             in_channels=channels,
             out_channels=channels,
@@ -353,7 +353,7 @@ class TestConvTranspose(flow.unittest.TestCase):
             stride=random() | nothing(),
             padding=random(1, 3).to(int) | nothing(),
             dilation=random(1, 5) | nothing(),
-            groups=random(1, 7),
+            groups=random(1, 6),
             padding_mode=constant("zeros") | nothing(),
         )
         m.train(random())