--- a/python/oneflow/test/expensive/test_einsum.py
+++ b/python/oneflow/test/expensive/test_einsum.py
@@ -22,14 +22,14 @@ from oneflow.test_utils.automated_test_util import *
 
 @flow.unittest.skip_unless_1n1d()
 class TestEinsum(flow.unittest.TestCase):
-    @autotest(n=20, check_graph=True)
+    @autotest(n=5)
     def test_einsum_matrix_transpose(test_case):
         device = random_device()
         x = random_tensor(ndim=2, dim0=random(1, 6), dim1=random(1, 6),).to(device)
         z = torch.einsum("ij->ji", x)
         return z
 
-    @autotest(n=20, check_graph=True)
+    @autotest(n=5)
     def test_einsum_eltwise_multiply(test_case):
         device = random_device()
         dim0 = random(1, 6)
@@ -39,7 +39,7 @@ class TestEinsum(flow.unittest.TestCase):
         z = torch.einsum("ij,ij->ij", x, y)
         return z
 
-    @autotest(n=20, check_graph=True)
+    @autotest(n=5)
     def test_einsum_get_diagonal(test_case):
         device = random_device()
         dim = random(1, 6)
@@ -47,7 +47,7 @@ class TestEinsum(flow.unittest.TestCase):
         z = torch.einsum("ii->i", x)
         return z
 
-    @autotest(n=20, check_graph=True)
+    @autotest(n=5)
     def test_einsum_batch_permute(test_case):
         device = random_device()
         x = random_tensor(
@@ -61,21 +61,21 @@ class TestEinsum(flow.unittest.TestCase):
         z = torch.einsum("...ij->...ji", x)
         return z
 
-    @autotest(n=20, check_graph=True)
+    @autotest(n=5)
     def test_einsum_reduce_sum(test_case):
         device = random_device()
         x = random_tensor(ndim=2, dim0=random(1, 6), dim1=random(1, 6),).to(device)
         z = torch.einsum("ij->", x)
         return z
 
-    @autotest(n=20, check_graph=True)
+    @autotest(n=5)
     def test_einsum_matrix_column_sum(test_case):
         device = random_device()
         x = random_tensor(ndim=2, dim0=random(1, 6), dim1=random(1, 6),).to(device)
         z = torch.einsum("ij->j", x)
         return z
 
-    @autotest(n=20, check_graph=True)
+    @autotest(n=5)
     def test_einsum_matrix_vector_multiply(test_case):
         device = random_device()
         dim0 = random(1, 6)
@@ -86,7 +86,7 @@ class TestEinsum(flow.unittest.TestCase):
         z = torch.einsum("ik,k", x, y)
         return z
 
-    @autotest(n=20, check_graph=True)
+    @autotest(n=5)
     def test_einsum_matmul(test_case):
         device = random_device()
         dim0 = random(1, 6)
@@ -98,7 +98,7 @@ class TestEinsum(flow.unittest.TestCase):
         z = torch.einsum("ik,kj", x, y)
         return z
 
-    @autotest(n=20, check_graph=True)
+    @autotest(n=5)
     def test_einsum_vector_inner_product(test_case):
         device = random_device()
         dim0 = random(1, 6)
@@ -108,7 +108,7 @@ class TestEinsum(flow.unittest.TestCase):
         z = torch.einsum("i,i", x, y)
         return z
 
-    @autotest(n=20, check_graph=True)
+    @autotest(n=5)
     def test_einsum_eltwise_mul_then_reduce_sum(test_case):
         device = random_device()
         dim0 = random(1, 6)
@@ -119,7 +119,7 @@ class TestEinsum(flow.unittest.TestCase):
         z = torch.einsum("ij,ij", x, y)
         return z
 
-    @autotest(n=20, check_graph=True)
+    @autotest(n=5)
     def test_einsum_vector_outer_product(test_case):
         device = random_device()
         x = random_tensor(ndim=1, dim0=random(1, 6),).to(device)
@@ -128,7 +128,7 @@ class TestEinsum(flow.unittest.TestCase):
         z = torch.einsum("i,j", x, y)
         return z
 
-    @autotest(n=20, check_graph=True)
+    @autotest(n=5)
     def test_einsum_batch_matmul(test_case):
         device = random_device()
         dim0 = random(1, 6)
@@ -138,7 +138,7 @@ class TestEinsum(flow.unittest.TestCase):
         z = torch.einsum("ijk,ikl->ijl", x, y)
         return z
 
-    @autotest(n=20, check_graph=True)
+    @autotest(n=5)
     def test_einsum_tensor_contraction(test_case):
         device = random_device()
         dim0 = random(1, 6)
@@ -157,7 +157,7 @@ class TestEinsum(flow.unittest.TestCase):
         z = torch.einsum("pqrs,tuqvr->pstuv", x, y)
         return z
 
-    @autotest(n=20, check_graph=True)
+    @autotest(n=5)
     def test_einsum_bilinear_transformation(test_case):
         device = random_device()
         dim0 = random(1, 6)
@@ -178,7 +178,7 @@ class TestEinsum(flow.unittest.TestCase):
         z = torch.einsum("ijk", x)
         return z
 
-    @autotest(n=20, check_graph=True)
+    @autotest(n=5)
     def test_einsum_tensor_contraction2(test_case):
         device = random_device()
         dim0 = random(1, 6)
@@ -189,7 +189,7 @@ class TestEinsum(flow.unittest.TestCase):
         z = torch.einsum("b n h w, n d -> b d h w", x, y)
         return z
 
-    @autotest(n=20, check_graph=True)
+    @autotest(n=5)
     def test_einsum_eltwise_mul_sum_row(test_case):
         device = random_device()
         dim0 = random(1, 6)
@@ -199,7 +199,7 @@ class TestEinsum(flow.unittest.TestCase):
         z = torch.einsum("n d, n d -> n", x, y)
         return z
 
-    @autotest(n=20, check_graph=True)
+    @autotest(n=5)
     def test_einsum_matmul2(test_case):
         device = random_device()
         dim0 = random(1, 6)
@@ -208,7 +208,7 @@ class TestEinsum(flow.unittest.TestCase):
         z = torch.einsum("i d, j d -> i j", x, y)
         return z
 
-    @autotest(n=20, check_graph=True)
+    @autotest(n=5)
     def test_einsum_attention(test_case):
         device = random_device()
         dim0 = random(1, 6)
@@ -223,7 +223,7 @@ class TestEinsum(flow.unittest.TestCase):
         z = torch.einsum("b h i d, b h j d -> b h i j", x, y)
         return z
 
-    @autotest(n=20, check_graph=True)
+    @autotest(n=5)
     def test_einsum_batch_matmul2(test_case):
         device = random_device()
         dim0 = random(1, 6)
@@ -238,7 +238,7 @@ class TestEinsum(flow.unittest.TestCase):
         z = torch.einsum("b h i j, b h j d -> b h i d", x, y)
         return z
 
-    @autotest(n=20, check_graph=True)
+    @autotest(n=5)
     def test_einsum_batch_matrix_vector_multiply(test_case):
         device = random_device()
         dim0 = random(1, 6)
@@ -251,7 +251,7 @@ class TestEinsum(flow.unittest.TestCase):
         z = torch.einsum("b i d, b i j d -> b i j", x, y)
         return z
 
-    @autotest(n=20, check_graph=True)
+    @autotest(n=5)
     def test_einsum_batch_matmul3(test_case):
         device = random_device()
         dim0 = random(1, 6)
@@ -263,7 +263,7 @@ class TestEinsum(flow.unittest.TestCase):
         z = torch.einsum("b x i d, b j d -> b x i j", x, y)
         return z
 
-    @autotest(n=20, check_graph=True)
+    @autotest(n=5)
     def test_einsum_batch_matmul4(test_case):
         device = random_device()
         dim0 = random(1, 6)
@@ -275,7 +275,7 @@ class TestEinsum(flow.unittest.TestCase):
         z = torch.einsum("b x i j, b j d -> b x i d", x, y)
         return z
 
-    @autotest(n=20, check_graph=True)
+    @autotest(n=5)
     def test_einsum_alphaflod_usecase1(test_case):
         device = random_device()
         dim0 = random(1, 6)
@@ -285,7 +285,7 @@ class TestEinsum(flow.unittest.TestCase):
         z = torch.einsum("hij, ijc->ihc", x, y)
         return z
 
-    @autotest(n=20, check_graph=True)
+    @autotest(n=5)
     def test_einsum_alphaflod_usecase2(test_case):
         device = random_device()
         dim0 = random(1, 6)
@@ -295,7 +295,7 @@ class TestEinsum(flow.unittest.TestCase):
         z = torch.einsum("rac,rab->rbc", x, y)
         return z
 
-    @autotest(n=20, check_graph=True)
+    @autotest(n=5)
     def test_einsum_alphaflod_usecase3(test_case):
         device = random_device()
         dim0 = random(1, 6)
@@ -305,7 +305,7 @@ class TestEinsum(flow.unittest.TestCase):
         z = torch.einsum("ra,rab->rb", x, y)
         return z
 
-    @autotest(n=20, check_graph=True)
+    @autotest(n=5)
     def test_einsum_alphaflod_usecase4(test_case):
         device = random_device()
         dim0 = random(1, 6)
@@ -315,7 +315,7 @@ class TestEinsum(flow.unittest.TestCase):
         z = torch.einsum("qhc,khc->qkh", x, y)
         return z
 
-    @autotest(n=20, check_graph=True)
+    @autotest(n=5)
     def test_einsum_alphaflod_usecase5(test_case):
         device = random_device()
         dim0 = random(1, 6)
@@ -326,7 +326,7 @@ class TestEinsum(flow.unittest.TestCase):
         z = torch.einsum("nm, mrc->nrc", x, y)
         return z
 
-    @autotest(n=20, check_graph=True)
+    @autotest(n=5)
     def test_einsum_alphaflod_usecase6(test_case):
         device = random_device()
         dim0 = random(1, 6)
@@ -336,7 +336,7 @@ class TestEinsum(flow.unittest.TestCase):
         z = torch.einsum("abc,adc->bdc", x, y)
         return z
 
-    @autotest(n=20, check_graph=True)
+    @autotest(n=5)
     def test_einsum_alphaflod_usecase7(test_case):
         device = random_device()
         dim0 = random(1, 6)
@@ -348,7 +348,7 @@ class TestEinsum(flow.unittest.TestCase):
         z = torch.einsum("dceb,cef->dbf", x, y)
         return z
 
-    @autotest(n=20, check_graph=True)
+    @autotest(n=5)
     def test_einsum_alphaflod_usecase8(test_case):
         device = random_device()
         dim0 = random(1, 6)
@@ -361,7 +361,7 @@ class TestEinsum(flow.unittest.TestCase):
         z = torch.einsum("acb,ade->dceb", x, y)
         return z
 
-    @autotest(n=20, check_graph=True)
+    @autotest(n=5)
     def test_einsum_alphaflod_usecase9(test_case):
         device = random_device()
         dim0 = random(1, 6)
@@ -372,7 +372,7 @@ class TestEinsum(flow.unittest.TestCase):
         z = torch.einsum("qkc,ch->hqk", x, y)
         return z
 
-    @autotest(n=20, check_graph=True)
+    @autotest(n=5)
     def test_einsum_alphaflod_usecase10(test_case):
         device = random_device()
         dim0 = random(1, 6)
@@ -387,7 +387,7 @@ class TestEinsum(flow.unittest.TestCase):
         z = torch.einsum("bhqk,bkhc->bqhc", x, y)
         return z
 
-    @autotest(n=20, check_graph=True)
+    @autotest(n=5)
     def test_einsum_alphaflod_usecase11(test_case):
         device = random_device()
         dim0 = random(1, 6)
@@ -400,7 +400,7 @@ class TestEinsum(flow.unittest.TestCase):
         z = torch.einsum("bqa,ahc->bqhc", x, y)
         return z
 
-    @autotest(n=20, check_graph=True)
+    @autotest(n=5)
     def test_einsum_ellipsis_usecase1(test_case):
         device = random_device()
         dim0 = random(1, 6)
@@ -413,7 +413,7 @@ class TestEinsum(flow.unittest.TestCase):
         z = torch.einsum("...lc, ...c -> ...l", x, y)
         return z
 
-    @autotest(n=20, check_graph=True)
+    @autotest(n=5)
     def test_einsum_ellipsis_usecase2(test_case):
         device = random_device()
         dim0 = random(1, 6)
@@ -423,7 +423,7 @@ class TestEinsum(flow.unittest.TestCase):
         z = torch.einsum("...lc, ...lc -> ...l", x, y)
         return z
 
-    @autotest(n=20, check_graph=True)
+    @autotest(n=5)
     def test_einsum_ellipsis_usecase3(test_case):
         device = random_device()
         dim0 = random(1, 6)
@@ -436,7 +436,7 @@ class TestEinsum(flow.unittest.TestCase):
         z = torch.einsum("...id,...jd->...ij", x, y)
         return z
 
-    @autotest(n=20, check_graph=True)
+    @autotest(n=5)
     def test_einsum_ellipsis_usecase4(test_case):
         device = random_device()
         dim0 = random(1, 6)
@@ -448,7 +448,7 @@ class TestEinsum(flow.unittest.TestCase):
         z = torch.einsum("...klm,kmn->...kln", x, y)
         return z
 
-    @autotest(n=20, check_graph=True)
+    @autotest(n=5)
     def test_einsum_ellipsis_usecase5(test_case):
         device = random_device()
         dim0 = random(1, 6)
@@ -461,7 +461,7 @@ class TestEinsum(flow.unittest.TestCase):
         z = torch.einsum("...ikl, ...jk -> ...ijl", x, y)
         return z
 
-    @autotest(n=20, check_graph=True)
+    @autotest(n=5)
     def test_einsum_ellipsis_usecase6(test_case):
         device = random_device()
         dim0 = random(1, 6)
@@ -474,7 +474,7 @@ class TestEinsum(flow.unittest.TestCase):
         z = torch.einsum("...l,...l->...", x, y)
         return z
 
-    @autotest(n=20, check_graph=True)
+    @autotest(n=5)
     def test_einsum_ellipsis_usecase7(test_case):
         device = random_device()
         dim0 = random(1, 6)
@@ -487,7 +487,7 @@ class TestEinsum(flow.unittest.TestCase):
         z = torch.einsum("ijk,ijk...->ij...", x, y)
         return z
 
-    @autotest(n=20, check_graph=True)
+    @autotest(n=5)
     def test_einsum_other_usecase1(test_case):
         device = random_device()
         dim0 = random(1, 6)
@@ -499,7 +499,7 @@ class TestEinsum(flow.unittest.TestCase):
         z = torch.einsum("bxi,oij,byj->boxy", x, y, w)
         return z
 
-    @autotest(n=20, check_graph=True)
+    @autotest(n=5)
     def test_einsum_other_usecase2(test_case):
         device = random_device()
         dim0 = random(1, 6)
@@ -513,7 +513,7 @@ class TestEinsum(flow.unittest.TestCase):
         z = torch.einsum("ijac,ijkp->ijakcp", x, y)
         return z
 
-    @autotest(n=20, check_graph=True)
+    @autotest(n=5)
     def test_einsum_other_usecase3(test_case):
         device = random_device()
         dim0 = random(1, 6)
@@ -525,7 +525,7 @@ class TestEinsum(flow.unittest.TestCase):
         z = torch.einsum("cdij,cbi->cdbj", x, y)
         return z
 
-    @autotest(n=20, check_graph=True)
+    @autotest(n=5)
     def test_einsum_fastfold_usecase1(test_case):
         device = random_device()
         dim0 = random(1, 6)
@@ -540,7 +540,7 @@ class TestEinsum(flow.unittest.TestCase):
         z = torch.einsum("bsid,bsjd->bijd", x, y)
         return z
 
-    @autotest(n=20, check_graph=True)
+    @autotest(n=5)
     def test_einsum_fastfold_usecase2(test_case):
         device = random_device()
         dim0 = random(1, 6)
@@ -554,7 +554,7 @@ class TestEinsum(flow.unittest.TestCase):
         z = torch.einsum("bsid,bsje->bijde", x, y)
         return z
 
-    @autotest(n=20, check_graph=True)
+    @autotest(n=5)
     def test_einsum_openfold_usecase1(test_case):
         device = random_device()
         dim0 = random(1, 6)
@@ -567,7 +567,7 @@ class TestEinsum(flow.unittest.TestCase):
         z = torch.einsum("...bac,...dae->...bdce", x, y)
         return z
 
-    @autotest(n=20, check_graph=True)
+    @autotest(n=5)
     def test_einsum_openfold_usecase2(test_case):
         device = random_device()
         dim0 = random(1, 6)
@@ -581,7 +581,7 @@ class TestEinsum(flow.unittest.TestCase):
         z = torch.einsum("...abc,...adc->...bdc", x, y)
         return z
 
-    @autotest(n=20, check_graph=True)
+    @autotest(n=5)
     def test_einsum_openfold_usecase3(test_case):
         device = random_device()
         dim0 = random(1, 6)
@@ -595,7 +595,7 @@ class TestEinsum(flow.unittest.TestCase):
         z = torch.einsum("...qhd,...khd->...hqk", x, y)
         return z
 
-    @autotest(n=20, check_graph=True)
+    @autotest(n=5)
     def test_einsum_openfold_usecase4(test_case):
         device = random_device()
         dim0 = random(1, 6)
@@ -609,7 +609,7 @@ class TestEinsum(flow.unittest.TestCase):
         z = torch.einsum("...vhf,...qhv->...qhf", x, y)
         return z
 
-    @autotest(n=20, check_graph=True)
+    @autotest(n=5)
     def test_einsum_openfold_usecase5(test_case):
         device = random_device()
         dim0 = random(1, 6)