--- a/python/oneflow/test/expensive/test_sqrt_square_sum.py
+++ b/python/oneflow/test/expensive/test_sqrt_square_sum.py
@@ -24,33 +24,32 @@ from oneflow.test_utils.automated_test_util import *
 
 @flow.unittest.skip_unless_1n1d()
 class TestLinalgVectorNorm2D(flow.unittest.TestCase):
-    @autotest(n=30, auto_backward=False, check_graph=True, rtol=0.5, atol=0.5)
+    @autotest(n=2, auto_backward=False, check_graph=True, rtol=0.5, atol=0.5)
     def test_sqrt_sum_with_cpu_random_data(test_case):
         device = cpu_device()
-        x = random_tensor(ndim=4, dim1=30, dim2=40, dim3=50, requires_grad=False).to(
+        x = random_tensor(ndim=4, dim1=3, dim2=4, dim3=5, requires_grad=False).to(
             device
         )
         y = torch.linalg.norm(x)
         return y
 
     @unittest.skipIf(os.getenv("ONEFLOW_TEST_CPU_ONLY"), "only test cpu cases")
-    @autotest(n=30, auto_backward=False, check_graph=True)
+    @autotest(n=2, auto_backward=False, check_graph=True)
     def test_sqrt_sum_with_cuda_random_data(test_case):
         device = gpu_device()
-        x = random_tensor(ndim=4, dim1=100, dim2=100, dim3=100, requires_grad=False).to(
+        x = random_tensor(ndim=4, dim1=10, dim2=10, dim3=10, requires_grad=False).to(
             device
         )
         y = torch.linalg.norm(x)
         return y
 
-    @autotest(n=30, auto_backward=False, check_graph=True, rtol=0.5, atol=0.5)
+    @autotest(n=2, auto_backward=False, check_graph=True, rtol=0.5, atol=0.5)
     def test_scalar_print_random_data(test_case):
         device = random_device()
-        x = random_tensor(ndim=4, dim1=30, dim2=40, dim3=50, requires_grad=False).to(
+        x = random_tensor(ndim=4, dim1=3, dim2=4, dim3=5, requires_grad=False).to(
             device
         )
         y = torch.linalg.norm(x)
-        print(f"grad_norm {y.oneflow:.4f}\t")
         return y